# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Digital Inspection App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inspections:
    get:
      operationId: listInspections
      tags: [inspections]
      summary: List inspections with filters
      parameters:
        - in: query
          name: outcome
          schema: { type: string, enum: [all, violation, no_violation] }
        - in: query
          name: categoryId
          schema: { type: integer }
        - in: query
          name: subCategoryId
          schema: { type: integer }
        - in: query
          name: severity
          schema: { type: string, enum: [A, B, C, D, E] }
        - in: query
          name: depotId
          schema: { type: integer }
        - in: query
          name: venueId
          schema: { type: integer }
        - in: query
          name: inspectorId
          schema: { type: integer }
        - in: query
          name: footageDateFrom
          schema: { type: string, format: date }
        - in: query
          name: footageDateTo
          schema: { type: string, format: date }
        - in: query
          name: inspectionDateFrom
          schema: { type: string, format: date }
        - in: query
          name: inspectionDateTo
          schema: { type: string, format: date }
        - in: query
          name: status
          schema: { type: string, enum: [in_progress, completed, all] }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: mine
          schema: { type: boolean }
      responses:
        "200":
          description: Inspections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InspectionSummary"
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Create an inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInspectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
  /inspections/{id}:
    get:
      operationId: getInspection
      tags: [inspections]
      summary: Get full inspection with findings
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
    patch:
      operationId: updateInspection
      tags: [inspections]
      summary: Update inspection header
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInspectionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
  /inspections/{id}/complete:
    post:
      operationId: completeInspection
      tags: [inspections]
      summary: Mark an inspection as complete
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    InspectionStatus:
      type: string
      enum: [in_progress, completed]
    CreateInspectionInput:
      type: object
      properties:
        dvrNumber: { type: string }
        depotId: { type: integer }
        venueId: { type: integer }
        driveId:
          type: ["integer", "null"]
        footageDate: { type: string, format: date }
        inspectionDate: { type: string, format: date-time }
        notes:
          type: ["string", "null"]
      required: [dvrNumber, depotId, venueId, footageDate]
    UpdateInspectionInput:
      type: object
      properties:
        dvrNumber: { type: string }
        depotId: { type: integer }
        venueId: { type: integer }
        footageDate: { type: string, format: date }
        inspectionDate: { type: string, format: date-time }
        notes:
          type: ["string", "null"]
    Severity:
      type: string
      enum: [A, B, C, D, E]
...
    Finding:
      type: object
      properties:
        id: { type: integer }
        inspectionId: { type: integer }
        clipName: { type: string }
        clipNumber: { type: integer }
        outcome: { type: string, enum: [no_violation, violation] }
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        subCategoryId:
          type: ["integer", "null"]
        subCategoryName:
          type: ["string", "null"]
        severity:
          oneOf:
            - $ref: "#/components/schemas/Severity"
            - type: "null"
        notes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
      required:
        - id
        - inspectionId
        - clipName
        - clipNumber
        - outcome
        - categoryId
        - categoryName
        - subCategoryId
        - subCategoryName
        - severity
        - notes
        - createdAt
    InspectionSummary:
      type: object
      properties:
        id: { type: integer }
        dvrNumber: { type: string }
        depotId: { type: integer }
        depotName: { type: string }
        venueId: { type: integer }
        venueName: { type: string }
        venueCode: { type: string }
        footageDate: { type: string, format: date }
        inspectionDate: { type: string, format: date-time }
        inspectorId: { type: integer }
        inspectorName: { type: string }
        driveId:
          type: ["integer", "null"]
        driveName:
          type: ["string", "null"]
        status: { $ref: "#/components/schemas/InspectionStatus" }
        findingsCount: { type: integer }
        violationsCount: { type: integer }
        highestSeverity:
          oneOf:
            - $ref: "#/components/schemas/Severity"
            - type: "null"
        createdAt: { type: string, format: date-time }
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - dvrNumber
        - depotId
        - depotName
        - venueId
        - venueName
        - venueCode
        - footageDate
        - inspectionDate
        - inspectorId
        - inspectorName
        - driveId
        - driveName
        - status
        - findingsCount
        - violationsCount
        - highestSeverity
        - createdAt
        - completedAt
    Inspection:
      type: object
      properties:
        id: { type: integer }
        dvrNumber: { type: string }
        depotId: { type: integer }
        depotName: { type: string }
        venueId: { type: integer }
        venueName: { type: string }
        venueCode: { type: string }
        footageDate: { type: string, format: date }
        inspectionDate: { type: string, format: date-time }
        inspectorId: { type: integer }
        inspectorName: { type: string }
        driveId:
          type: ["integer", "null"]
        driveName:
          type: ["string", "null"]
        status: { $ref: "#/components/schemas/InspectionStatus" }
        notes:
          type: ["string", "null"]
        findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        createdAt: { type: string, format: date-time }
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - dvrNumber
        - depotId
        - depotName
        - venueId
        - venueName
        - venueCode
        - footageDate
        - inspectionDate
        - inspectorId
        - inspectorName
        - driveId
        - driveName
        - status
        - notes
        - findings
        - createdAt
        - completedAt
...