# lib/api-zod/src/generated/types/inspection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Inspection App API
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from "./finding";
import type { InspectionStatus } from "./inspectionStatus";

export interface Inspection {
  id: number;
  dvrNumber: string;
  depotId: number;
  depotName: string;
  venueId: number;
  venueName: string;
  venueCode: string;
  footageDate: Date;
  inspectionDate: Date;
  inspectorId: number;
  inspectorName: string;
  /** @nullable */
  driveId: number | null;
  /** @nullable */
  driveName: string | null;
  status: InspectionStatus;
  /** @nullable */
  notes: string | null;
  findings: Finding[];
  createdAt: Date;
  /** @nullable */
  completedAt: Date | null;
}